Доработка ДЗ по комментариям от наставника 09,04
