+++ hwDraftComments_Модульный_JS/assistants.js
@@ -1,4 +1,6 @@
-export  const currentDate = new Date();
+const buttonElement = document.getElementById ('add-button');
+
+export const currentDate = new Date();
 function formatTime(currentDate){
 
  const date = currentDate.getDate().toString().padStart(2, "0");
@@ -10,3 +12,28 @@ function formatTime(currentDate){
 
  return `${date}.${month}.${year} ${h}:${m}`
 };
+
+export function delay(interval = 300) {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve();
+      }, interval);
+    });
+};
+
+const formElement = document.getElementById ('form');
+ formElement.addEventListener('keyup', keyEvent);
+  function keyEvent(e) {
+    if (e.code === 'Enter') {
+     buttonElement.dispatchEvent(new Event('click'));
+    }
+  };
+ 
+const removeButton = document.getElementById("deleteComment");
+removeButton.addEventListener("click", () => {
+  сommentators.pop();
+  renderCommentators();
+});
+
+
+
