+++ hwDraftComments_Модульный_JS/main.js
@@ -1,9 +1,5 @@
 import { getComments, getPost  } from "./api.js";
-import {сommentators, renderCommentators, setComments } from "./renderCommentators.js";
-
-window.onload = function() {
-  let preloader = document.getElementById('preloader');
-};
+import { setComments, renderCommentators } from "./renderCommentators.js";
 
 const nameElement = document.getElementById ('name-input');
 const textElement = document.getElementById ('text-input');
@@ -11,6 +7,9 @@ const buttonElement = document.getElementById ('add-button');
 const addForm = document.getElementById("form");
 const loader = document.querySelector(".loader");
 
+window.onload = function() {
+  let preloader = document.getElementById('preloader');
+};
 
 function getCom() {
   getComments().then((responseData) => {
@@ -25,37 +24,14 @@ function getCom() {
       };
     });
 
-    setComments(appComments);
-    renderCommentators();
-    loader.textContent = '';
-    addForm.classList.remove("hidden");
-    preloader.classList.add('preloader-hidden');
+   setComments(appComments);
+   renderCommentators();
+   loader.textContent = '';
+   addForm.classList.remove("hidden");
+   preloader.classList.add('preloader-hidden');
 
   });
 }
-
-export function editEventListeners () {
-  const editButtonElements = document.querySelectorAll(".edit-button"); 
-  for ( const editButtonEl of editButtonElements) { 
-    editButtonEl.addEventListener("click", (e) => { 
-      e.stopPropagation();
-      const index = editButtonEl.dataset.index;
-      сommentators[index].isEdit = !сommentators[index].isEdit;
-      renderCommentators();
-    });
-  }
-
-  const saveButtons = document.querySelectorAll(".save-button");
-  for (const saveButton of saveButtons) {
-    saveButton.addEventListener("click", (e) => {
-      e.stopPropagation();
-      const index = saveButton.dataset.index;
-      сommentators[index].isEdit = !сommentators[index].isEdit;
-      сommentators[index].comment = saveButton.closest('.comment').querySelector('textarea').value
-      renderCommentators();
-    });
-  }
-};
 
 getCom();
 
@@ -84,26 +60,12 @@ buttonElement.addEventListener("click", () => {
     textElement.value = "";
     return getCom();
   })
-  // console.warn(error);
-  loader.textContent = '';
   addForm.classList.remove("hidden");
+  loader.textContent = '';
+
 });
 
 renderCommentators();
-
-const formElement = document.getElementById ('form');
- formElement.addEventListener('keyup', keyEvent);
-  function keyEvent(e) {
-    if (e.code === 'Enter') {
-     buttonElement.dispatchEvent(new Event('click'));
-    }
-  };
- 
-const removeButton = document.getElementById("deleteComment");
-removeButton.addEventListener("click", () => {
-  сommentators.pop();
-  renderCommentators();
-});
 
 console.log("It works!");
