+++ hwDraftComments_Модульный_JS/renderCommentators.js
@@ -1,7 +1,10 @@
-import { editEventListeners } from "./main.js";
+import { delay } from "./assistants.js";
+import { sanitizeHtml } from "./validation.js";
 
 const listElement = document.getElementById ('list');
 const textElement = document.getElementById ('text-input');
+
+
 
 export let сommentators = [];
 
@@ -13,11 +16,11 @@ export const renderCommentators = () => {
   const commentatorsHtml = сommentators.map((сommentator, index) => {
     return `<li class="comment" data-index="${index}">
       <div class="comment-header">
-       <div>${сommentator.name}</div>
+       <div>${sanitizeHtml(сommentator.name)}</div>
        <div>${сommentator.date}</div>
       </div>
       <div class="comment-body">
-       <div class="${сommentator.isEdit ? 'display-none' : 'comment-text'}">${сommentator.comment}</div>
+       <div class="${сommentator.isEdit ? 'display-none' : 'comment-text'}">${sanitizeHtml(сommentator.comment)}</div>
        <textarea type="textarea" class="${сommentator.isEdit ? 'add-form-text' : 'display-none'}" >${сommentator.comment}</textarea>
       </div>
       <div class="comment-footer">
@@ -39,23 +42,15 @@ export const renderCommentators = () => {
   answerComment();
 };
 
-function delay(interval = 300) {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve();
-    }, interval);
-  });
-}   
-
-const initEventListeners = () => {
+let initEventListeners = () => {
   const likeButtonElement = document.querySelectorAll(".like-button");
   for ( const likeButtonEl of likeButtonElement) {
 
     const index = likeButtonEl.dataset.index;
     const counter = likeButtonEl.dataset.like;
 
-    likeButtonEl.addEventListener("click", (event) => {
-      event.stopPropagation();
+    likeButtonEl.addEventListener("click", (e) => {
+      e.stopPropagation();
 
       likeButtonEl.classList.add("-loading-like");
 
@@ -83,6 +78,29 @@ const initEventListeners = () => {
 }
 renderCommentators();
 
+function editEventListeners () {
+  const editButtonElements = document.querySelectorAll(".edit-button"); 
+  for ( const editButtonEl of editButtonElements) { 
+    editButtonEl.addEventListener("click", (e) => { 
+      e.stopPropagation();
+      const index = editButtonEl.dataset.index;
+      сommentators[index].isEdit = !сommentators[index].isEdit;
+      renderCommentators();
+    });
+  }
+
+  const saveButtons = document.querySelectorAll(".save-button");
+  for (const saveButton of saveButtons) {
+    saveButton.addEventListener("click", (e) => {
+      e.stopPropagation();
+      const index = saveButton.dataset.index;
+      сommentators[index].isEdit = !сommentators[index].isEdit;
+      сommentators[index].comment = saveButton.closest('.comment').querySelector('textarea').value
+      renderCommentators();
+    });
+  }
+};
+
 function answerComment () {
       
   const commentsElements = document.querySelectorAll('.comment');
@@ -102,14 +120,5 @@ function answerComment () {
   }
 };
 
-export const sanitizeHtml = (htmlString) => {
-  return htmlString
-  .replaceAll("&", "&amp;")
-  .replaceAll("<", "&lt;")
-  .replaceAll(">", "&gt;")
-  .replaceAll('"', "&quot;")
-  .replaceAll("QUOTE_BEGIN", "<div class='quote'>")
-  .replaceAll("QUOTE_END", "</div>")
-};
