+++ hwDraftComments_Модульный_JS/validation.js
@@ -1,0 +1,9 @@
+export const sanitizeHtml = (htmlString) => {
+    return htmlString
+    .replaceAll("&", "&amp;")
+    .replaceAll("<", "&lt;")
+    .replaceAll(">", "&gt;")
+    .replaceAll('"', "&quot;")
+    .replaceAll("QUOTE_BEGIN", "<div class='quote'>")
+    .replaceAll("QUOTE_END", "</div>")
+};
